+++ jaxley/modules/network.py
@@ -148,6 +148,7 @@ class Network(Module):
         )
         self._solve_indexer = JaxleySolveIndexer(
             cumsum_ncomp=padded_cumsum_ncomp,
+            ncomp_per_branch=self.ncomp_per_branch,
             branchpoint_group_inds=branchpoint_group_inds,
             children_in_level=children_in_level,
             parents_in_level=parents_in_level,