+++ CHANGELOG.md
@@ -1,11 +1,34 @@
-# 0.6.3
+# 0.7.0
+
+### New Features
+
+- Allow ion diffusion with `cell.diffuse()` and add tutorials (#438, @michaeldeistler):
+```python
+from jaxley.channels import CaNernstReversal
+from jaxley.pumps import CaFaradayConcentrationChange
+
+cell.insert(CaFaradayConcentrationChange())
+cell.insert(CaNernstReversal())
+cell.diffuse("CaCon_i")
+cell.set("axial_diffusion_CaCon_i", 1.0)
+```
+- Introduce ion pumps (#438, @michaeldeistler)
+
+### Bug fixes
+
+- Fix for simulation of morphologies with inhomogenous numbers of compartments (#438, @michaeldeistler)
+- Bugfix for types assigned by SWC reader when soma is traced by a single point (#582, @Kartik-Sama, @michaeldeistler).
+
+### Code health
 
 - add new release workflow (#588, @jnsbck)
-- Bugfix for types assigned by SWC reader when soma is traced by a single point (#582, @Kartik-Sama, @michaeldeistler).
+- update FAQ and tutorials (#593, @michaeldeistler)
+
 
 # 0.6.2
 
 - also remove `nseg` from tutorials (#580, @jnsbck)
+
 
 # 0.6.1